dir: revert "dir: select directories correctly"
This reverts commit 28aeec4380d7ff1306d9c61f37d670eb0a3f7a61.

The change in 28aeec4 (dir: select directories correctly, 2021-09-24)
caused a regression in directory-based matches with non-cone-mode
patterns, especially for .gitignore patterns. A test is included to
prevent this regression in the future.

The commit e541ea16 (dir: fix pattern matching on dirs, 2021-09-24) was
reverted in 2fe88d4 (dir: fix directory-matching bug, 2021-11-02) for
similar reasons. Neither commit changed tests, and tests added later in
the series continue to pass when these commits are reverted.

Reported-by: Danial Alihosseini <danial.alihosseini@gmail.com>
Signed-off-by: Derrick Stolee <dstolee@microsoft.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>